Update patch set 1
Patch Set 1: Code-Review+1

I deserialized the json-ld for relation queries to always return relationspans (sorted). When would a krill query ask only for relation sources/targets?

I implemented relations with attribute, but it seems that Koral does not support attributes in a relation object. So, I haven't implement the deserialization for this.

Patch-set: 1
Label: Code-Review=+2