Update patch set 1
Patch Set 1:

> Is it common to describe running instructions in pom? I think it
 > would be better in the readme.

Of course, if we had public corpora available, it would make sense to have it in the readme. Describing the build cycles in the pom however makes sense. My point was rather: It's now wrong what is described in the pom.xml and it has lost its's formatting.

Patch-set: 1
Label: Code-Review=+2