Update patch set 2
Patch Set 2:

> Is the cache loaded in case only the indexer is started?

Yes. Please see KrillCollection line #84. Does the indexer not need any KrillCollection? Alternatively, the cache could be first loaded in the instructors. I suppose the cache is only accessed with a KrillCollection instantiation, e.g. in a search, except when clearing the cache, e.g. in commit.

Patch-set: 2